refactor: reorganize structure of js modules
