+++ js/api/create-comment-data.js
@@ -1,4 +1,4 @@
-const createComment = (actions) => {
+const createCommentData = (actions) => {
   const {getId, getAvatar, getMessage, getName} = actions;
 
   const id = getId();
@@ -9,4 +9,4 @@ const createComment = (actions) => {
   return {id, avatar, message, name};
 };
 
-export {createComment};
+export {createCommentData};
