+++ js/api/get-mock-data.js
@@ -1,0 +1,6 @@
+import { getCommentsData } from './get-comments-data.js';
+import { getPhotosData } from './get-photos-data.js';
+
+const getMockData = () => getPhotosData(getCommentsData);
+
+export {getMockData};
