+++ js/api/index.js
@@ -1,0 +1,1 @@
+export { getMockData } from './get-mock-data.js';
