--- js/data.js
@@ -1,5 +1,0 @@
-import { getPhotosData } from './photos-data.js';
-
-const getMockData = () => getPhotosData();
-
-export {getMockData};
