+++ js/thumbnails/create-thumbnail.js
