+++ js/thumbnails/index.js
@@ -1,0 +1,2 @@
+export { getThumbnails } from './get-thumbnails.js';
+export { renderThumbnails } from './render-thumbnails.js';
