+++ js/thumbnails/render-thumbnails.js
@@ -1,0 +1,7 @@
+import { toOuterHTML } from '../util';
+
+const picturesContainer = document.querySelector('.pictures');
+
+const renderThumbnails = (thumbnailElems) => picturesContainer.insertAdjacentHTML('beforeend', toOuterHTML(thumbnailElems));
+
+export {renderThumbnails};
